cleanup(core): rename repo-key to workspace-id and require an explicit root
deriveRepoKey no longer defaults its directory to workspaceRoot - callers pass it, so the module holds no implicit global state. generateWorkspaceId moves here from analytics-prompt (which is about consent prompting, not identity) and likewise takes an explicit root. The new git helpers use execFileSync rather than execSync: the commands are constant today, but an argv array keeps a later edit from interpolating a variable into a shell string, and it skips the shell entirely.
